version 3.0.2

+++ frontEnd/js/main.js
@@ -1,7 +1,8 @@
 const notes = document.querySelector(".todo-heading");
-
 let deleteNote = document.getElementById("deleteNote");
 const api = "http://localhost:8002/notes";
+let ids;
+
 const getAllNotes = () => {
   fetch("http://localhost:8002/notes/get")
     .then((response) => response.json())
@@ -10,13 +11,14 @@ const getAllNotes = () => {
       document.getElementById("display").innerHTML = "";
       for (let i = 0; i < json.data.length; i++) {
         const apiId = json.data[i]._id;
-        // console.log(apiId);
         document.getElementById(
           "display"
-        ).innerHTML += `<div class="cards col-5  text-center  ">
+        ).innerHTML += `<div class="cards col-10 col-md-3  text-center  ">
            <div class="display_title ">${json.data[i].title}</div>
           <div class="text-start d-flex  justify-content-between align-items-center  text-wrap"><div>${json.data[i].content}</div>
-          <button class="btn btn-light my-3 col-2 p-2" id="deleteNote " onclick="del('${apiId}')"><img src="../trash.svg" alt=""></button>
+         <div class="d-flex flex-column  col-3 gap-2">
+         <button class="btn del_btn btn-light" id="deleteNote" onclick="del('${apiId}')"><i class="fa-solid fa-trash"></i></button>
+          <button class="btn del_btn btn-light  " id="editNote" onclick="getNoteById('${apiId}')"><i class="fa-regular fa-pen-to-square"></i></i></button></div>
           </div></div>`;
         document.getElementById("title").value = "";
         document.getElementById("dec").value = "";
@@ -24,13 +26,18 @@ const getAllNotes = () => {
     });
 };
 
-getAllNotes();
-
 function del(i) {
-  fetch(`http://localhost:8002/notes/delete/${i}`, {
-    method: "DELETE",
-  });
-  getAllNotes();
+  if (confirm("Are you sure to delete this note??") == true) {
+    fetch(`http://localhost:8002/notes/delete/${i}`, {
+      method: "DELETE",
+    })
+      .then((response) => response.json())
+      .then((json) => {
+        console.log(json);
+        document.getElementById("message").innerHTML = `${json.message}`;
+      });
+    getAllNotes();
+  }
 }
 
 function add() {
@@ -60,11 +67,55 @@ function add() {
           <div class="text-start d-flex flex-column " id="final_desc">${
             document.getElementById("dec").value
           }</div>
-          <button class="btn btn-light my-3 p-2" ><img src="../trash.svg" alt=""></button>
+          <button class="btn btn-light my-3 p-2" ><i class="fa-solid fa-trash"></i></button>
       </div>`;
+        document.getElementById("message").innerHTML = `${json.message}`;
         document.getElementById("title").value = "";
         document.getElementById("dec").value = "";
         getAllNotes();
       });
   }
 }
+
+function getNoteById(i) {
+  fetch(`http://localhost:8002/notes/${i}`)
+    .then((response) => response.json())
+    .then((json) => {
+      console.log(json);
+      const notesID = json.data._id;
+      const title = json.data.title;
+      const desc = json.data.content;
+      console.log(title);
+      update(notesID,title,desc);
+    });
+}
+function update(id,title,desc) {
+  document.getElementById("enter").classList.add("hide");
+  document.getElementById("update").classList.remove("hide");
+  document.getElementById("title").value=title;
+  document.getElementById("dec").value=desc;
+  document
+    .getElementById("update")
+    .addEventListener("click", async function () {
+      await fetch(`http://localhost:8002/notes/update/${id}`, {
+        method: "PUT",
+        body: JSON.stringify({
+          title: document.getElementById("title").value,
+          content: document.getElementById("dec").value,
+        }),
+        headers: {
+          "Content-type": "application/json; charset=UTF-8",
+        },
+      }) .then((response) => response.json())
+      .then((json) => {
+        document.getElementById("message").innerHTML = `${json.message}`;
+      });
+      document.getElementById("title").value = "";
+      document.getElementById("dec").value = "";
+      getAllNotes();
+      document.getElementById("enter").classList.remove("hide");
+      document.getElementById("update").classList.add("hide");
+    });
+}
+
+getAllNotes();
